Major revision of command line argument parse construction.
Instead of 100s of argparse function calls to setup parsers and
sub parsers and add arguments, etc., a single datastructure at the
top of the file was creating that contains the parse tree as lists
of dictionarys with lists, etc.  The user interface map can very
quickly be rewritten by modifying the data structure instead of
lots of lines of argparse function calls.